fix: ts migration
