+++ src/components/search/highlighted-text.tsx
@@ -1,0 +1,32 @@
+/**
+ * Copyright (c) 2024, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+import { FC } from 'react';
+
+interface HighlightedTextProps {
+    text: string;
+    highlight: string;
+}
+
+export const HighlightedText: FC<HighlightedTextProps> = (props) => {
+    const { text, highlight } = props;
+    const parts = text.split(new RegExp(`(${highlight})`, 'gi'));
+
+    return (
+        <span>
+            {parts.map((part, i) =>
+                part.toLowerCase() === highlight.toLowerCase() ? (
+                    <span key={i} style={{ fontWeight: 'bold' }}>
+                        {part}
+                    </span>
+                ) : (
+                    part
+                )
+            )}
+        </span>
+    );
+};
